requirements.txt
# Client-side only: portable arrow-key selector (client.py prompt_move).
# Install with: pip install -r requirements.txt
questionary>=2.0
pyyaml>=6.0.1
